requirements.txt
Flask==3.0.0
reportlab==4.0.7
Pillow==10.1.0
requests==2.31.0
